requirements.txt
google-generativeai
dateparser
